requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
click==8.3.1
colorama==0.4.6
decorator==5.2.1
distro==1.9.0
fastapi==0.135.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
jiter==0.13.0
moviepy==2.2.1
numpy==2.4.2
openai==2.26.0
opencv-python==4.13.0.92
pillow==11.3.0
proglog==0.1.12
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
sniffio==1.3.1
starlette==0.52.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
